+++ config/strict.js
@@ -1,0 +1,21 @@
+export default {
+  rules: {
+    "carbon/layout-use": [true, { enforceScopes: true, severity: "error" }],
+    "carbon/motion-duration-use": [
+      true,
+      { enforceScopes: true, severity: "error" }
+    ],
+    "carbon/motion-easing-use": [
+      true,
+      { enforceScopes: true, severity: "error" }
+    ],
+    "carbon/theme-use": [true, { enforceScopes: true, severity: "error" }],
+    "carbon/type-use": [
+      true,
+      {
+        enforceScopes: true,
+        severity: "error"
+      }
+    ]
+  }
+};