+++ src/rules/layout-use/__tests__/index.js
@@ -224,322 +224,28 @@ testRule({
     {
       code: ".foo { right: $spacing-01; }",
       description: 'A right using a token "$spacing-01" is accepted.'
-    }
-  ],
-  reject: [
+    },
     {
       code: ".foo { margin: $container-01; }",
-      description: 'A margin using non-token(s) "$container-01" is rejected',
-      message: messages.rejectedUndefinedVariable("margin", "$container-01")
+      description: 'A margin using token(s) "$container-01" is accepted'
     },
     {
       code: ".foo { margin: $container-01 $fluid-spacing-01; }",
       description:
-        'A margin using non-token(s) "$container-01 $fluid-spacing-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        }
-      ]
+        'A margin using token(s) "$container-01 $fluid-spacing-01" is accepted'
     },
     {
       code: ".foo { margin: $container-01 $fluid-spacing-01 $icon-size-01; }",
       description:
-        'A margin using non-token(s) "$container-01 $fluid-spacing-01 $icon-size-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        }
-      ]
+        'A margin using token(s) "$container-01 $fluid-spacing-01 $icon-size-01" is accepted'
     },
     {
       code: ".foo { margin: $container-01 $fluid-spacing-01 $icon-size-01 $container-01; }",
       description:
-        'A margin using non-token(s) "$container-01 $fluid-spacing-01 $icon-size-01 $container-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $fluid-spacing-01; }",
-      description:
-        'A margin using non-token(s) "$fluid-spacing-01" is rejected',
-      message: messages.rejectedUndefinedVariable("margin", "$fluid-spacing-01")
-    },
-    {
-      code: ".foo { margin: $fluid-spacing-01 $icon-size-01; }",
-      description:
-        'A margin using non-token(s) "$fluid-spacing-01 $icon-size-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $fluid-spacing-01 $icon-size-01 $container-01; }",
-      description:
-        'A margin using non-token(s) "$fluid-spacing-01 $icon-size-01 $container-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $fluid-spacing-01 $icon-size-01 $container-01 $fluid-spacing-01; }",
-      description:
-        'A margin using non-token(s) "$fluid-spacing-01 $icon-size-01 $container-01 $fluid-spacing-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $icon-size-01; }",
-      description: 'A margin using non-token(s) "$icon-size-01" is rejected',
-      message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-    },
-    {
-      code: ".foo { margin: $icon-size-01 $container-01; }",
-      description:
-        'A margin using non-token(s) "$icon-size-01 $container-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $icon-size-01 $container-01 $fluid-spacing-01; }",
-      description:
-        'A margin using non-token(s) "$icon-size-01 $container-01 $fluid-spacing-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $icon-size-01 $container-01 $fluid-spacing-01 $icon-size-01; }",
-      description:
-        'A margin using non-token(s) "$icon-size-01 $container-01 $fluid-spacing-01 $icon-size-01" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $container-01 $fluid-spacing-01 $icon-size-01 199px; }",
-      description:
-        'A margin using non-token(s) "$container-01 $fluid-spacing-01 $icon-size-01 199px" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$container-01")
-        },
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejected(
-            "margin",
-            "$container-01 $fluid-spacing-01 $icon-size-01 199px"
-          )
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $fluid-spacing-01 $icon-size-01 199px; }",
-      description:
-        'A margin using non-token(s) "$fluid-spacing-01 $icon-size-01 199px" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejected(
-            "margin",
-            "$fluid-spacing-01 $icon-size-01 199px"
-          )
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $fluid-spacing-01 $icon-size-01 199px 299px; }",
-      description:
-        'A margin using non-token(s) "$fluid-spacing-01 $icon-size-01 199px 299px" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable(
-            "margin",
-            "$fluid-spacing-01"
-          )
-        },
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejected(
-            "margin",
-            "$fluid-spacing-01 $icon-size-01 199px 299px"
-          )
-        },
-        {
-          message: messages.rejected(
-            "margin",
-            "$fluid-spacing-01 $icon-size-01 199px 299px"
-          )
-        }
-      ]
-    },
-    {
-      code: ".foo { margin: $icon-size-01 199px; }",
-      description:
-        'A margin using non-token(s) "$icon-size-01 199px" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        { message: messages.rejected("margin", "$icon-size-01 199px") }
-      ]
-    },
-    {
-      code: ".foo { margin: $icon-size-01 199px 299px; }",
-      description:
-        'A margin using non-token(s) "$icon-size-01 199px 299px" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        { message: messages.rejected("margin", "$icon-size-01 199px 299px") },
-        { message: messages.rejected("margin", "$icon-size-01 199px 299px") }
-      ]
-    },
-    {
-      code: ".foo { margin: $icon-size-01 199px 299px 399px; }",
-      description:
-        'A margin using non-token(s) "$icon-size-01 199px 299px 399px" is rejected',
-      warnings: [
-        {
-          message: messages.rejectedUndefinedVariable("margin", "$icon-size-01")
-        },
-        {
-          message: messages.rejected(
-            "margin",
-            "$icon-size-01 199px 299px 399px"
-          )
-        },
-        {
-          message: messages.rejected(
-            "margin",
-            "$icon-size-01 199px 299px 399px"
-          )
-        },
-        {
-          message: messages.rejected(
-            "margin",
-            "$icon-size-01 199px 299px 399px"
-          )
-        }
-      ]
-    },
+        'A margin using token(s) "$container-01 $fluid-spacing-01 $icon-size-01 $container-01" is accepted'
+    }
+  ],
+  reject: [
     {
       code: ".foo { margin: 199px; }",
       description: 'A margin using non-token(s) "199px" is rejected',
@@ -627,26 +333,6 @@ testRule({
           )
         }
       ]
-    },
-    {
-      code: ".foo { left: $container-01; }",
-      description: 'A left using a non-token "$container-01" is rejected.',
-      message: messages.rejectedUndefinedVariable("left", "$container-01")
-    },
-    {
-      code: ".foo { top: $container-01; }",
-      description: 'A top using a non-token "$container-01" is rejected.',
-      message: messages.rejectedUndefinedVariable("top", "$container-01")
-    },
-    {
-      code: ".foo { bottom: $container-01; }",
-      description: 'A bottom using a non-token "$container-01" is rejected.',
-      message: messages.rejectedUndefinedVariable("bottom", "$container-01")
-    },
-    {
-      code: ".foo { right: $container-01; }",
-      description: 'A right using a non-token "$container-01" is rejected.',
-      message: messages.rejectedUndefinedVariable("right", "$container-01")
     }
   ]
 });
@@ -655,20 +341,11 @@ testRule({
   plugins: [plugin],
   ruleName,
   customSyntax: "postcss-scss",
-  config: [
-    true,
-    {
-      acceptContainerTokens: true
-    }
-  ],
+  config: [true],
   accept: [
     {
       code: `.foo { left: $container-01; }`,
-      description: `Accept $container tokens with acceptContainerTokens: true.`
-    },
-    {
-      code: `.foo { left: $container-01; }`,
-      description: `Accept $container tokens with acceptContainerTokens: true.`
+      description: `Accept $container tokens`
     }
   ],
   reject: []
@@ -678,21 +355,12 @@ testRule({
   plugins: [plugin],
   ruleName,
   customSyntax: "postcss-scss",
-  config: [
-    true,
-    {
-      acceptIconSizeTokens: true
-    }
-  ],
+  config: [true],
 
   accept: [
     {
       code: `.foo { left: $icon-size-01; }`,
-      description: `Accept $icon-size tokens with acceptIconSizeTokens: true.`
-    },
-    {
-      code: `.foo { left: $icon-size-01; }`,
-      description: `Accept $icon-size tokens with acceptIconSizeTokens: true.`
+      description: `Accept $icon-size tokens`
     }
   ],
   reject: []
@@ -702,21 +370,12 @@ testRule({
   plugins: [plugin],
   ruleName,
   customSyntax: "postcss-scss",
-  config: [
-    true,
-    {
-      acceptFluidSpacingTokens: true
-    }
-  ],
+  config: [true],
 
   accept: [
     {
       code: `.foo { left: $fluid-spacing-01; }`,
-      description: `Accept $fluid-spacing tokens with acceptFluidSpacingTokens: true.`
-    },
-    {
-      code: `.foo { left: $fluid-spacing-01; }`,
-      description: `Accept $fluid-spacing tokens with acceptFluidSpacingTokens: true.`
+      description: `Accept $fluid-spacing tokens`
     }
   ],
   reject: []
@@ -789,14 +448,7 @@ testRule({
   ruleName,
   customSyntax: "postcss-scss",
   fix: true,
-  config: [
-    true,
-    {
-      acceptContainerTokens: true,
-      acceptFluidSpacingTokens: true,
-      acceptIconSizeTokens: true
-    }
-  ],
+  config: [true],
   reject: [
     {
       code: `.foo { left: $carbon--layout-04; }`,
@@ -1865,6 +1517,35 @@ testRule({
       code: `.foo { padding-block-end: 10px }`,
       description: "padding-block-end using tokens",
       message: messages.rejected("padding-block-end", "10px")
+    }
+  ]
+});
+
+testRule({
+  plugins: [plugin],
+  ruleName,
+  config: [true],
+  configSyntax: "postcss-scss",
+  accept: [
+    {
+      code: `.foo { gap: $spacing-01 $spacing-01 }`,
+      description: "gap using tokens"
+    },
+    {
+      code: `.foo { gap: $spacing-01}`,
+      description: "gap using single token"
+    }
+  ],
+  reject: [
+    {
+      code: `.foo { gap: 10px $spacing-01 }`,
+      description: "gap not using token in list",
+      message: messages.rejected("gap", "10px $spacing-01")
+    },
+    {
+      code: `.foo { gap: 10px }`,
+      description: "gap not using a token",
+      message: messages.rejected("gap", "10px")
     }
   ]
 });