+++ src/rules/type-use/utils/initTypeTokens.js
@@ -36,39 +36,31 @@ const doInit = async ({ carbonPath, carbonModulePostfix }) => {
   if (isV10) {
     typeFunctions = [
       {
-        name: "carbon--font-style",
-        accept: "acceptCarbonFontStyleFunction"
-      },
-      {
         name: "carbon--font-weight",
-        accept: "acceptCarbonFontWeightFunction"
+        accept: true
       },
       {
         name: "carbon--type-scale",
-        accept: "acceptCarbonTypeScaleFunction"
+        accept: true
       },
       {
         name: "carbon--font-family",
-        accept: "acceptCarbonFontFamilyFunction"
+        accept: true
       }
     ];
   } else {
     typeFunctions = [
       {
-        name: "font-style",
-        accept: "acceptCarbonFontStyleFunction"
-      },
-      {
         name: "font-weight",
-        accept: "acceptCarbonFontWeightFunction"
+        accept: true
       },
       {
         name: "type-scale",
-        accept: "acceptCarbonTypeScaleFunction"
+        accept: true
       },
       {
         name: "font-family",
-        accept: "acceptCarbonFontFamilyFunction"
+        accept: true
       }
     ];
   }